Unique analyze file for a4 with arguments
